backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
sqlalchemy
